Rename evaluation.tests.test_utils to tools
