+++ evap/contributor/tests/test_forms.py
@@ -4,7 +4,7 @@ from model_mommy import mommy
 
 from evap.evaluation.models import UserProfile, Course, Questionnaire, Contribution
 from evap.contributor.forms import DelegatesForm, EditorContributionForm
-from evap.evaluation.tests.test_utils import WebTest, get_form_data_from_instance, to_querydict
+from evap.evaluation.tests.tools import WebTest, get_form_data_from_instance, to_querydict
 from evap.staff.forms import ContributionFormSet
